directly return boolean using the evaluation statement, prevent returning undefined: null object pattern
